fixed post request v2 command, added a http request base to inherit from, and mypy and ruff are now passing w/ burnettk
