+++ src/connector_http/commands/post_request.py
@@ -1,9 +1,11 @@
 from typing import Any
 
-import requests
+import requests  # type: ignore
+from spiffworkflow_connector_command.command_interface import CommandResultDictV1
+from spiffworkflow_connector_command.command_interface import ConnectorCommand
 
 
-class PostRequest:
+class PostRequest(ConnectorCommand):
     def __init__(self,
         url: str,
         headers: dict[str, str] | None,
@@ -17,7 +19,7 @@ class PostRequest:
         self.basic_auth_password = basic_auth_password
         self.data = data
 
-    def execute(self, config, task_data):
+    def execute(self, _config: Any, _task_data: Any) -> CommandResultDictV1:
         auth = None
         if self.basic_auth_username is not None and self.basic_auth_password is not None:
             auth = (self.basic_auth_username, self.basic_auth_password)
