+++ tests/connector_http/unit/test_post_request_v2.py
@@ -2,19 +2,20 @@ import json
 from typing import Any
 from unittest.mock import patch
 
-from connector_http.commands.get_request_v2 import GetRequestV2
+from connector_http.commands.post_request_v2 import PostRequestV2
 
 
-class TestGetRequestV2:
+class TestPostRequestV2:
     def test_html_from_url(self) -> None:
-        request = GetRequestV2(url="http://example.com")
+        request = PostRequestV2(url="http://example.com")
         result = None
         return_html = "<html>Hey</html>"
-        with patch("requests.get") as mock_request:
+        with patch("requests.post") as mock_request:
             mock_request.return_value.status_code = 200
             mock_request.return_value.ok = True
             mock_request.return_value.text = return_html
             result = request.execute(None, {})
+            assert mock_request.call_count == 1
         assert result is not None
         assert result["status"] == 200
         assert result["mimetype"] == "application/json"
@@ -27,15 +28,16 @@ class TestGetRequestV2:
         assert len(response["spiff__logs"]) > 0
 
     def test_json_from_url(self) -> None:
-        request = GetRequestV2(url="http://example.com")
+        request = PostRequestV2(url="http://example.com")
         result = None
         return_json = {"hey": "we_return"}
-        with patch("requests.get") as mock_request:
+        with patch("requests.post") as mock_request:
             mock_request.return_value.status_code = 200
             mock_request.return_value.ok = True
             mock_request.return_value.headers = {"Content-Type": "application/json"}
             mock_request.return_value.text = json.dumps(return_json)
             result = request.execute(None, {})
+            assert mock_request.call_count == 1
         assert result is not None
         assert result["status"] == 200
         assert result["mimetype"] == "application/json"
@@ -48,15 +50,15 @@ class TestGetRequestV2:
         assert len(response["spiff__logs"]) > 0
 
     def test_can_handle_500(self, sleepless: Any) -> None:
-        request = GetRequestV2(url="http://example.com", attempts=3)
+        request = PostRequestV2(url="http://example.com")
         result = None
         return_json = {"error": "we_did_error"}
-        with patch("requests.get") as mock_request:
+        with patch("requests.post") as mock_request:
             mock_request.return_value.status_code = 500
             mock_request.return_value.headers = {"Content-Type": "application/json"}
             mock_request.return_value.text = json.dumps(return_json)
             result = request.execute(None, {})
-            assert mock_request.call_count == 3
+            assert mock_request.call_count == 1
         assert result is not None
         assert result["status"] == 500
         assert result["mimetype"] == "application/json"
